refactor: move the email things to a separate file
